requirements.txt
autopep8==1.5.7
big-O-calculator==0.0.9.8.4
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mpmath==1.2.1
pycodestyle==2.7.0
sympy==1.8
toml==0.10.2
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.5.0
